+++ changelog.md
@@ -8,6 +8,10 @@ Now shows the equip location of recently awarded items.
 
 Added new chat command `/rc export` which will export a csv formatted list of the items currently in session.
 
+### Button Groups
+
+Added button groups for mounts, bags and recipies.
+
 ## Bugfixes
 
 - *Items can now again be automatically added to a trade with a player from another realm.*